fix: address code quality review issues
- Frame.__hash__ = None instead of raising TypeError (Pythonic pattern)
- Add @functools.total_ordering to ID3TimeStamp for complete ordering
- Add Frame.__eq__ based on HashKey comparison
- Add v2.2 frame aliases to id3.py __all__ for proper re-export
- Wrap MP4 covr values in MP4Cover, freeform values in MP4FreeForm
- Fix all 14 clippy warnings (Default derives, collapsed ifs, etc.)

+++ src/flac/mod.rs
@@ -82,8 +82,8 @@ impl StreamInfo {
         let sr_hi = u32::from_be_bytes([0, data[10], data[11], data[12]]);
         let sample_rate = (sr_hi >> 4) & 0xFFFFF;
 
-        let channels = (((data[12] >> 1) & 0x07) + 1) as u8;
-        let bps_hi = ((data[12] & 0x01) as u8) << 4;
+        let channels = ((data[12] >> 1) & 0x07) + 1 ;
+        let bps_hi = (data[12] & 0x01) << 4;
         let bps_lo = (data[13] >> 4) & 0x0F;
         let bits_per_sample = (bps_hi | bps_lo) + 1;
